+++ src/app/components/activity/activity-list.component.ts
@@ -1,11 +1,13 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Inject } from '@angular/core';
 import { FirebaseService } from '../../services/firebase.service';
 import { Router } from '@angular/router';
 import * as moment from 'moment';
 import { Activity } from 'src/app/models/activity';
 import { DatePipe } from '@angular/common';
+import {MatDialog, MatDialogRef, MAT_DIALOG_DATA} from '@angular/material/dialog';
 
 import {ExcelService} from '../../services/excel.service';
+import { DialogOkCancelData, DialogOkCancelComponent } from '../dialog/dialog-ok-cancel.component';
 
 @Component({
     selector: 'app-activity-list',
@@ -14,15 +16,17 @@ import {ExcelService} from '../../services/excel.service';
 })
 export class ActivityListComponent implements OnInit {
     COLLECTION = 'activities';
+    COLLECTION_USER = 'users';
     datePipe = new DatePipe('en-US');
     activities: Array<any>;
-
+    users: {};
 
     /* Constructor */
     constructor(
         private excelService: ExcelService,
         private router: Router,
-        public firebaseService: FirebaseService
+        public firebaseService: FirebaseService,
+        public dialog: MatDialog
     ) {
         // Set localStorage: currentEntity.
         localStorage.setItem('currentEntity', 'activity');
@@ -30,10 +34,21 @@ export class ActivityListComponent implements OnInit {
     /* OnInit */
     ngOnInit() {
         this.getActivities();
+        this.getUsers();
     }
     /*----------------------------- */
     /*---------- Methods ---------- */
     /*----------------------------- */
+
+    getUsers() {
+        this.firebaseService.getDocuments(this.COLLECTION_USER).subscribe(result => {
+            const users = {};
+            result.forEach(function(item) {
+                users[item.payload.doc.id] = item.payload.doc.data();
+            });
+            this.users = users;
+        });
+    }
 
     getActivities() {
         const lastModifiedBy = localStorage.getItem('idUser');
@@ -45,17 +60,45 @@ export class ActivityListComponent implements OnInit {
                 }
                 return activity;
             });
+            const datePipe = this.datePipe;
+            // Sort by workDate.
+            this.activities = this.activities.sort(function(item1: any, item2: any){
+                const value1 = datePipe.transform(item1.workDate, 'yyyyMMdd').toLowerCase();
+                const value2 = datePipe.transform(item2.workDate, 'yyyyMMdd').toLowerCase();
+                if (value1 > value2) { return 1; }
+                if (value1 < value2) { return -1; }
+                return 0;
+            });
         });
     }
 
     deleteActivity(id) {
-        this.firebaseService.deleteDocument(this.COLLECTION, id).then(result => {
-            alert('Deleted');
+        const dialogData: DialogOkCancelData = { title: 'Warning', content: 'Are you sure to delect it?', result: -1 };
+        const dialogRef = this.dialog.open(DialogOkCancelComponent, {
+            data: dialogData
+        });
+
+        dialogRef.afterClosed().subscribe(result => {
+            if (dialogData.result === 1) {
+                this.firebaseService.deleteDocument(this.COLLECTION, id).then(result => {
+                });
+            }
         });
     }
 
     
-    exportAsXLSX():void {
-        this.excelService.exportAsExcelFile(this.activities, 'DailyReport_' + this.datePipe.transform(new Date(), 'yyyyMMdd'), 'daily');
+    exportAsXLSX(): void {
+        const excelData = this.activities.map(item => {
+            const data = {};
+            data['Project Name'] = item.projectName;
+            data['Summary'] = item.name;
+            data['Type'] = item.type;
+            data['Start Date'] = item.workDate;
+            data['Assignee'] = this.users[item.lastModifiedBy].name;
+            data['Report To'] = item.reportTo;
+            data['Status'] = item.status;
+            return data;
+        });
+        this.excelService.exportAsExcelFile(excelData, 'DailyReport_' + this.datePipe.transform(new Date(), 'yyyyMMdd'), 'daily');
     }
 }
